chore: split obvious class into files
